Renamed android package to `com.dooboolab.app`
