+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.dooboolab">
+  package="com.dooboolab.app">
 
     <uses-permission android:name="android.permission.INTERNET" />
